+++ lib/model/constants.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+const inactiveColor = Colors.amber;
+const activeColor = Colors.red;
