+++ lib/model/container_color.dart
@@ -1,0 +1,4 @@
+enum ContainerColor {
+  first,
+  second,
+}
